Fix for App Breaking "Alphabet" is read-only error
Current implementation breaks application when using AtoZList Component. By setting the Alphabet variable to a mutable variable rather than a constant (on line 16 of AlphabetPicker.js) resolves this error. If this pull request gets accepted please publish the fix to NPM as well, so that this becomes a working component again for applications that rely on it.